shifted out analyzeSolution, analyzeSolutions and analyzeSolutionsForPublicPaths
