+++ functions/index.js
@@ -12,7 +12,9 @@ const jupyterTrigger = require("./src/executeJupyterSolution");
 const downloadEvents = require("./src/downloadEvents");
 const solutionTriggers = require("./src/updateSolutionVisibility");
 const httpUtil= require("./src/utils/http").httpUtil;
-const updateSkills = require("./src/updateSkills");
+const pubPathSol = require("./src/analyzeSolutionsForPublicPaths");
+const singleSol = require("./src/analyzeSolution");
+
 
 const profilesRefreshApproach =
   (functions.config().profiles &&
@@ -142,132 +144,12 @@ exports.updateFPP = functions.https.onRequest((request, response) => {
 
 exports.analyzeSolutionGivenUserAndProbKey = functions.https.onRequest((req, res) => {
 
-  // function to format data into desired dictionary:
-  const formatData = (qns, user, data, result) => {
-    // console.log(data["cells"][0]["source"])
-    result[qns] = {}
-    result[qns][user] = {}
-    result[qns][user] = data["cells"][0]["source"]
-    return result
-  }
-
-  // function to retrieve "problemSolutions" node from database:
-  async function getSpecificSolutions (probkey, userkey) {
-    var rawSingleSol = null;
-    const specificRoute = admin.database().ref(`problemSolutions/${probkey}/${userkey}/`).once("value").then(snapshot => {
-      // console.log(snapshot.val())
-      rawSingleSol = snapshot.val();
-    })
-    await specificRoute
-    return rawSingleSol;
-  }
-
-  // create a dictionary to store output
-  var reformattedSingleSol = {}
-
-  // async function to do await
-  async function reformatData(probkey, userkey) {
-    var rawSingleSolPromise = await getSpecificSolutions(probkey, userkey);
-    const formatDataPromise = formatData(probkey, userkey, rawSingleSolPromise, reformattedSingleSol);
-    await formatDataPromise;
-    console.log(reformattedSingleSol)
-  }
-
-  reformatData("-LEgrNOmnWsFkr6nFHwO", "eVJVC9kde3QSiXAP989kivD9SZn2")
+  singleSol.analyzeSolution("-LEgrNOmnWsFkr6nFHwO", "eVJVC9kde3QSiXAP989kivD9SZn2")
 
   res.status(200).send("FUNCTION ONE ERROR-FREE~ ^^");
 });
 
 exports.analyzeSolutionsForPublicPaths = functions.https.onRequest((request, response) => {
-  // variable to store a list of all public paths:
-  var allPublicPaths = []
-
-  // 1. extract out all public paths:
-  async function getPubPaths() {
-    const pathRoute = admin.database().ref("paths/").once("value").then(snapshot => {
-      const bigPathDic = snapshot.val();
-      for (var pathkey in bigPathDic) {
-        var pathInfo = bigPathDic[pathkey];
-        if(pathInfo.hasOwnProperty("isPublic")) {
-          allPublicPaths.push(pathkey);
-        }
-      }
-    })
-    await pathRoute
-    // console.log(allPublicPaths)
-    return allPublicPaths
-  }
-
-  // getPubPaths()
-
-  // variable to store all problems that belong to public paths:
-  var allPublicPathProblems = []
-
-  // 2. extract out all problems that belong to public paths:
-  async function getPubQns(pubPathList) {
-    const qnsRoute = admin.database().ref("problems/").once("value").then(snapshot => {
-      const bigQnsDic = snapshot.val();
-      for (var user in bigQnsDic) {
-        var probDic = bigQnsDic[user];
-        for (var prob in probDic) {
-          var probInfo = probDic[prob];
-          if (probInfo.hasOwnProperty("path")) {
-            pathDetail = probInfo["path"];
-            if(pubPathList.includes(pathDetail)) {
-              allPublicPathProblems.push(prob);
-            }
-          }
-        }
-      }
-    })
-    await qnsRoute
-    // console.log(allPublicPathProblems)
-    return allPublicPathProblems
-  }
-
-  // Create a dictionary to store solution dictionary
-  var solForPublicProblems = {}
-
-  // 3. retrieve all the solutions to the public qns
-  async function getSolutions(pubQnsList) {
-    const qnsRoute = admin.database().ref("problemSolutions/").once("value").then(snapshot => {
-      var bigSolDic = snapshot.val();
-      for (var probkey in bigSolDic) {
-        if (pubQnsList.includes(probkey)) {
-          var userSolDic = bigSolDic[probkey];
-          for (var userkey in userSolDic) {
-            var usersol = userSolDic[userkey];
-            if (usersol.hasOwnProperty("cells")) {
-              // if problemkey not in dic, create new entry
-              if (!(probkey in solForPublicProblems) ) {
-                solForPublicProblems[probkey] = {}
-              }
-              solForPublicProblems[probkey][userkey] = usersol["cells"][0]["source"];
-            }
-          }
-        }
-      }
-    })
-    await qnsRoute
-    // console.log(solForPublicProblems)
-    return solForPublicProblems
-  }
-
-  // an async function that automatically gets called to run all the above functions
-  (async() => {
-    const pubPath = await getPubPaths();
-    const pubQns = await getPubQns(pubPath);
-    // console.log(pubQns)
-    const pubSol = await getSolutions(pubQns);
-    console.log(solForPublicProblems)
-  }) ()
-
+  pubPathSol.analyzeSolutionsForPublicPaths();
   response.status(200).send("PART TWO DONE :D")
-})
-
-
-exports.yrtest = functions.https.onRequest((request, response) => {
-  updateSkills.updateBothSkills({"problemSkills": {"problems": {"oneSkillType": {"skillFeature": {"sausage": true}}}}, "userSkills": {"cherry": {"function": {"next": {"c": true}}}}})
-
-  response.status(200).send("YR TEST SUCCESS :DD")
 })
