+++ functions/src/analyzeSolutions.js
